fix c11

+++ C11/ex05/include/operator.h
@@ -6,7 +6,7 @@
 /*   By: sooykim <sooykim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/22 02:34:16 by sooykim           #+#    #+#             */
-/*   Updated: 2021/09/25 01:31:20 by sooykim          ###   ########.fr       */
+/*   Updated: 2021/09/29 01:51:10 by sooykim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@
 # define DIV 3
 # define MOD 4
 
-long long int	add(int a, int b);
-long long int	sub(int a, int b);
-long long int	mul(int a, int b);
-long long int	div(int a, int b);
-long long int	mod(int a, int b);
+long long int	add(long long int a, long long int b);
+long long int	sub(long long int a, long long int b);
+long long int	mul(long long int a, long long int b);
+long long int	div(long long int a, long long int b);
+long long int	mod(long long int a, long long int b);
 
 #endif
