+++ C11/ex05/include/parser.h
@@ -6,7 +6,7 @@
 /*   By: sooykim <sooykim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/22 02:37:47 by sooykim           #+#    #+#             */
-/*   Updated: 2021/09/25 01:08:48 by sooykim          ###   ########.fr       */
+/*   Updated: 2021/09/29 01:34:57 by sooykim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@
 # define DIV 3
 # define MOD 4
 
-int	parse_int(char *a);
-int	classifyop(char op);
-int	parse_op(char *op);
+long long int	parse_int(char *a);
+int				classifyop(char op);
+int				parse_op(char *op);
 
 #endif
